feat: separates chart rendering to a component aside of the container
