+++ src/Features/MeasuresChart/MeasuresChart.styles.ts
@@ -2,7 +2,7 @@ import { makeStyles, Theme, createStyles } from '@material-ui/core/styles';
 
 export const useStyles = makeStyles((theme: Theme) => createStyles({
   paper: {
-    padding: theme.spacing(2),
+    padding: theme.spacing(1),
     marginBottom: theme.spacing(3),
   },
 }));
